finland: remove tezagm
The network connection at Tezagm's current home is being discontinued
and it's moving to my family which is behind CGN and often without IPv6
connectivity, so I have to remove it from here.

I expect it to continue working as a Tor peer when it's connected at its
new home.