Update united-states.md
Added the State of Missouri, which is where Kansas City (mostly) is, and added our Kansas City peer. The other KC peer is likely actually in Kansas City, Missouri as well but didn't want to make an adjustment to that listing based on an assumption.

Added the state of Pennsylvania, and added our Allentown, PA peer.

Added our Liberty Lake, Washington peer.